# doc/openapi.yaml
openapi: 3.0.0
info:
  title: Simple e-learning platform API
  version: 1.0.0
servers:
  # - url: https://inquisitive-outerwear.cyclic.app/
  #   description: Production server
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
paths:
  /auth/register:
    post:
      tags:
        - auth
      summary: regist
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                username: author2
                password: '123456'
                email: johnz@example.com
                role: author
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
        '400':
          description: Bad request
          content:
            application/json: {}
        '500' : 
          description: Internal Server Error
          content:
            application/json: {}
  /auth/login:
    post:
      tags:
        - auth
      summary: login
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                username: admin
                password: '123456'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
        '400':
          description: Bad request
          content:
            application/json: {}
        '500' : 
          description: Internal Server Error
          content:
            application/json: {}
...